backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
langchain==0.0.350
openai==1.3.0
pydantic==2.4.2
pydantic-settings==2.0.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
tiktoken==0.5.1
arxiv==2.1.0
beautifulsoup4==4.12.2
requests==2.31.0
pytest==7.4.2
httpx==0.25.0
numpy==1.26.0
pandas==2.1.1
matplotlib==3.8.0
seaborn==0.13.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
